TECH-345: Migrate TabItem to typescript
