+++ src/components/Tab/TabItem/variants/_ghost.scss
