+++ src/components/index.ts
@@ -19,6 +19,6 @@ export * from './SecondaryNav';
 export * from './Separator';
 export * from './SvgWrapper';
 export * from './Tab';
-export * from './TabItem';
+export * from './Tab/TabItem';
 export * from './TreeView';
 export * from './Typography';
